15 out of 27 from defining classes complete. :turtle:
- Imroved code in every exercise in this chapter.

+++ chapter-14-defining-classes/prgm-1401.cs
@@ -28,7 +28,7 @@ namespace Program
         string fullName;
         string email;
         int course;
-        int phoneNumber;
+        long phoneNumber;
         enum Subjects { Math, Physics, Philosophy }
         private Subjects subject;
         enum Universities { Standford, IBM, Harvard }
