+++ chapter-14-defining-classes/prgm-1402.cs
@@ -30,7 +30,7 @@ namespace Program
     {
         private string fullName = null;
         private string email = null;
-        private int phoneNumber = 0;
+        private long phoneNumber = 0;
         private int course = 0;
         public enum Subjects { Math, Physics, Philosophy }
         private Subjects subject;
@@ -49,10 +49,10 @@ namespace Program
         public Student(string fullName, string email)
             : this(fullName, email, 0) { }
 
-        public Student(string fullName, string email, int phoneNumber)
+        public Student(string fullName, string email, long phoneNumber)
             : this(fullName, email, phoneNumber, 0, Subjects.Math, Universities.Harvard) { }
 
-        public Student(string fullName, string email, int phoneNumber, int course, Subjects subject, Universities university)
+        public Student(string fullName, string email, long phoneNumber, int course, Subjects subject, Universities university)
         {
             this.fullName = fullName;
             this.email = email;
